change to UseEffect to retry
